aimu_lsm6ds33 compiling and running without error. Ready for test with hardware.

+++ apps/aimu_lsm6ds33/fsw/src/aimu_lsm6ds33_msg.h
@@ -70,12 +70,12 @@ typedef struct
 typedef struct 
 {
     uint8               TlmHeader[CFE_SB_TLM_HDR_SIZE];
-    uint16              AIMU_LSM6DS33_ACCELERATIONX;
-    uint16              AIMU_LSM6DS33_ACCELERATIONY;
-    uint16              AIMU_LSM6DS33_ACCELERATIONZ;
-    uint16              AIMU_LSM6DS33_ANGULAR_RATEX;
-    uint16              AIMU_LSM6DS33_ANGULAR_RATEY;
-    uint16              AIMU_LSM6DS33_ANGULAR_RATEZ;
+    float              AIMU_LSM6DS33_ACCELERATIONX;
+    float              AIMU_LSM6DS33_ACCELERATIONY;
+    float              AIMU_LSM6DS33_ACCELERATIONZ;
+    float              AIMU_LSM6DS33_ANGULAR_RATEX;
+    float              AIMU_LSM6DS33_ANGULAR_RATEY;
+    float              AIMU_LSM6DS33_ANGULAR_RATEZ;
 
 }   OS_PACK aimu_lsm6ds33_data_tlm_t  ;
